feat: now we can edit a dream

+++ app/src/main/java/com/example/dreamary/views/activities/Dreams/AddDreamActivity.kt
@@ -16,7 +16,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
@@ -63,42 +62,31 @@ import kotlinx.coroutines.launch
 import android.Manifest
 import android.app.TimePickerDialog
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.gestures.detectHorizontalDragGestures
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.material3.Card
 import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.OutlinedTextFieldDefaults
 import androidx.compose.material3.Slider
 import androidx.compose.material3.SliderDefaults
 import com.google.accompanist.flowlayout.FlowRow
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.KeyboardCapitalization
-import androidx.compose.ui.unit.sp
 import androidx.compose.ui.zIndex
-import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
-import com.bumptech.glide.integration.compose.GlideImage
 import com.example.dreamary.models.entities.Tag
 import com.example.dreamary.viewmodels.audio.AudioRecorderViewModel
 import com.example.dreamary.viewmodels.audio.AudioRecorderViewModelFactory
 import com.example.dreamary.views.components.CustomDropdown
 import java.util.Calendar
-import androidx.compose.ui.text.input.ImeAction
 import coil.compose.AsyncImage
 import com.example.dreamary.models.routes.NavRoutes
 import com.example.dreamary.views.components.DreamTextFieldCustom
 import com.example.dreamary.views.components.Loading
-import kotlin.math.abs
 
 @Preview(showBackground = true)
 @Composable
@@ -552,7 +540,7 @@ fun AddDreamActivity (navController: NavController, viewModel: AddDreamViewModel
                         ContextSleep(
                             noiseLevel = noiseLevel,
                             nbReveils = nbReveils,
-                            temperature = temperature,
+                            temperature = temperature.toLong(),
                             time = time,
                             onNoiseLevelChanged = { noiseLevel = it },
                             onNbReveilsChanged = { nbReveils = it },
@@ -802,10 +790,10 @@ fun showTimePicker(context: Context, onTimeSelected: (Int, Int) -> Unit) {
 
 @Composable
 fun ContextSleep (
-    noiseLevel : String,
-    nbReveils : String,
-    temperature : Int,
-    time : String,
+    noiseLevel: String,
+    nbReveils: String,
+    temperature: Long,
+    time: String,
     onNoiseLevelChanged: (String) -> Unit,
     onNbReveilsChanged: (String) -> Unit,
     onTemperatureChanged: (Int) -> Unit,
